requirements.txt
google-generativeai
faiss-cpu
beautifulsoup4
nltk
numpy
selenium
streamlit